requirements.txt
fastapi==0.115.12
mangum==0.19.0
pydantic-settings==2.9.1
python-binance==1.0.22
uvicorn==0.34.2
